Skeleton for model for the game
GameBoard holds GameTiles that hold GamePieces
In trying to separate the view from the model, the BiColorPiece uses a
boolean to represent color. This way you can see most importantly, is
any given piece of one team or another, and then the coloring is left to
the view. For GameTile there isn't necessarily any reason for only two
colors to exist so I'm going to try color name strings for options.