Changes to GameBoard and drafting model manager
Due to how private stuff works it makes more sense to have GameBoard be
an interface and allow a new class CheckersBoard to do checkers specific
things like the checkers pattern. 
Working out how to design the model manager. Created Location class to
act as a bridge for transferring data to the view.