Work on rule-check model
Begun attempt at validating moves

I had previously forgotten about the existence of King Pieces.
BiColorPieces worked fine for a checker board because there wasn't any
way to differentiate between pieces of the same team, but now that there
can be a king piece an extension to BiColorPiece is necessary.